requirements.txt
fastapi[standard]
requests
firebase-admin
python-dotenv